Fixed some issues with tex indices (can be negative if not set in obj) in obj loader

+++ RayTracer/ObjLoader.cpp
@@ -332,15 +332,15 @@ bool ObjLoader::Load(const std::string& name, bool center)
 		// also it's worth looking into "Optimal convex decompositions" by Bernard Chazelle and David Dobkin - a concave polygon can be split into convex ones
 
 		const size_t indexvertex1 = std::get<0>(polygon[startPoint]);
-		const size_t indextex1 = std::get<1>(polygon[startPoint]);
+		const long long int indextex1 = std::get<1>(polygon[startPoint]);
 		const size_t indexnormal1 = std::get<2>(polygon[startPoint]);
 
 		const size_t indexvertex2 = std::get<0>(polygon[(startPoint + 1ULL) % polygon.size()]);
-		const size_t indextex2 = std::get<1>(polygon[(startPoint + 1ULL) % polygon.size()]);
+		const long long int indextex2 = std::get<1>(polygon[(startPoint + 1ULL) % polygon.size()]);
 		const size_t indexnormal2 = std::get<2>(polygon[(startPoint + 1ULL) % polygon.size()]);
 
 		const size_t indexvertex3 = std::get<0>(polygon[(startPoint + 2ULL) % polygon.size()]);
-		const size_t indextex3 = std::get<1>(polygon[(startPoint + 2ULL) % polygon.size()]);
+		const long long int indextex3 = std::get<1>(polygon[(startPoint + 2ULL) % polygon.size()]);
 		const size_t indexnormal3 = std::get<2>(polygon[(startPoint + 2ULL) % polygon.size()]);
 
 		if (indexvertex1 >= vertices.size()) break;
@@ -352,11 +352,11 @@ bool ObjLoader::Load(const std::string& name, bool center)
 		if (indexnormal3 >= normals.size()) break;
 
 		const Vector3D<double> firstPoint = vertices[indexvertex1];
-		const size_t firstIndexTex = indextex1;
+		const long long int firstIndexTex = indextex1;
 		const Vector3D<double> firstNormal = normals[indexnormal1];
 
 		Vector3D<double> lastPoint(vertices[indexvertex3]);
-		size_t lastIndexTex = indextex3;
+		long long int lastIndexTex = indextex3;
 		Vector3D<double> lastNormal(normals[indexnormal3]);
 
 		std::shared_ptr<Objects::Triangle> triangle = std::make_shared<Objects::Triangle>(firstPoint, vertices[indexvertex2], lastPoint, firstNormal, normals[indexnormal2], lastNormal, material);
@@ -411,7 +411,7 @@ bool ObjLoader::Load(const std::string& name, bool center)
 			if (nextIndexNormal >= normals.size())
 				break;
 
-			const size_t nextIndexTex = std::get<1>(polygon[ind]);
+			const long long int nextIndexTex = std::get<1>(polygon[ind]);
 
 			const Vector3D<double>& nextPoint = vertices[nextIndexVertex];
 			const Vector3D<double>& nextNormal = normals[nextIndexNormal];
